+++ CHANGELOG.md
@@ -76,6 +76,7 @@ NOTE: This is just a convenience table to increase the readability of the config
 
 ### Bug Fix
 
+- Fix annotation comment tooltip going off screen
 - Fix color setting in the wrong format (ARGB instead of RGBA)
 - Fix not able to save after adding comment
 - Fix search results not clearing when searching for empty string